prototype of game to use for task reward
game where you are a blue square that dodges red rectangles of death.
Game is in 1 minute bouts. Rewards is given in seconds of being able to
shoot bullets that destroy the red rectangles. Punishment is 30 seconds
of not being able to play the game but having to press enter every 5
seconds so you can't go do something else.